Moved utility components to nested folder; better org.
